Safe Handler Execution
Evaluating 'safe' or 'apiSafe' in a 'filters' block allows the user to mark
certain messages for safe handling. When marked thus, any matching handlers
for these messages are evaluated without fully dequeueing the message from
the priority queue, such that if we receive and handle (i.e. swallow) an
exit signal, the message will be available for re-processing. Only once a
message has been safely processed will it be fully dequeued from the internal
mailbox/queue.